requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
fonttools==4.56.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pygame==2.6.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
scipy==1.15.2
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
